full_audio/requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
pillow
unitree-webrtc-connect
